+++ source/orcustrators/adventureOrcustrator.js
@@ -491,7 +491,7 @@ function resolveMove(move, adventure) {
  * @param {Adventure} adventure
  */
 function decrementDurability(moveName, user, adventure) {
-	if (!["Punch", "Appease"].includes(moveName) && user.team === "delver") {
+	if (!["Punch", "Appease", "Greed"].includes(moveName) && user.team === "delver") {
 		const gearCategory = getGearProperty(moveName, "category");
 		if (gearCategory === "Weapon") {
 			const weaponPolishCount = adventure.getArtifactCount("Weapon Polish");