+++ source/rooms/artifactguardian-treasureelemental.js
@@ -8,5 +8,5 @@ module.exports = new RoomTemplate("A windfall of treasure!",
 		new ResourceTemplate("3", "internal", "levelsGained"),
 		new ResourceTemplate("1", "loot", "artifact")
 	],
-	generateCombatRoomBuilder([])
+	generateCombatRoomBuilder(["greed"])
 ).addEnemy("Treasure Elemental", "1");